+++ CHANGELOG.md
@@ -16,7 +16,9 @@ for further information about branching and tagging conventions.
 - Resolved an issue when participant email doesn't show on `Participants` tab
   on `Poll` page
 - Resolved an issue when user can't update sale request
-
+- Resolved an issue when in select field quote assets on the atomic swap form
+ show only assets in user balance
+ 
 ## [1.10.0-x.0] - 2019-07-02
 #### Added
 - New Atomic swap feature: